Feature: Videos (#1)
## Description
Adds support to Videos to the site. Images or videos can now be added to the new `WorkEntry` content model, which is referenced in a `WorkPage`. This enables showing a carousel of both images and videos. Currently videos are set to autoplay and do no have a poster image. The videos are not lazily loaded, and will start playing as soon as the site loads.

This PR also adds Contentful migration functionality, and a migration to convert all existing images on the `WorkPage` to the `WorkEntry` type.

This change also adds a `yarn.lock` file.